txconfirm: encapsulate fee-input fanout in a controller FSM
The fee-input fanout logic sprawled across two files and ~23 functions,
with its state (the in-flight fanout, the per-parent predicted outputs)
living as loose fields on CPFPBroadcaster. The broadcaster, the actor,
and the fanout all reached into each other's state, so the contract
between txconfirm and the fanout was wide and hard to follow.

This pulls the whole mechanism behind a FeeBumpInputController. The
at-most-one in-flight fanout is now modeled as an in-memory protofsm
(idle <-> fanout-pending), mirroring the tracked-tx FSM: transitions are
pure, and the actor stays the IO driver, registering confirmation
watches and running broadcasts around the controller. The actor now
speaks to the controller through a narrow surface -- EnsureSupply,
OnFanoutConfirmed, PruneParent, PendingFanout -- instead of poking
broadcaster internals.

The reservation map (parentStates) keeps a single owner: the controller
holds a broadcaster reference and shares it, so the CPFP-child build
path and the fanout's predicted->used promotion stay consistent.
Behavior is unchanged; all txconfirm tests still pass.
